+++ tests/git_utils_test.py
@@ -18,6 +18,15 @@ def test_parse_git_remote_url_https():
     )
 
 
+def test_parse_git_remote_url_username_password():
+    assert _parse_git_remote_url(
+        "https://USERNAME:PASSWORD@github.com/allenai/beaker-gantry.git"
+    ) == (
+        "allenai",
+        "beaker-gantry",
+    )
+
+
 def test_invalid_git_remote_url():
     with pytest.raises(InvalidRemoteError):
-        _parse_git_remote_url("git@github.com/allenai/beaker-gantry.git")
+        _parse_git_remote_url("git@gthub.com:allenai/beaker-gantry.git")